backend/requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.5.3
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-allauth==0.61.1
django-crontab==0.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==24.14.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.2.0
msal==1.28.0
oauthlib==3.2.2
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
